Refactor reconciliation logic
