+++ src/main/java/com/alertops/AlertOpsApplication.java
@@ -3,10 +3,8 @@ package com.alertops;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cache.annotation.EnableCaching;
-//import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication
-//@EnableScheduling
 @EnableCaching
 public class AlertOpsApplication {
 	public static void main(String[] args) {
