--- src/main/java/com/alertops/flow_execution_engine/component/AppConfig.java
@@ -1,17 +1,0 @@
-package com.alertops.flow_execution_engine.component;
-/*
-package com.alertops.component;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-
-@Component
-public class AppConfig {
-    @Value("${app.default-role:user}")
-    private String defaultRole;
-
-    public String getDefaultRole() {
-        return defaultRole;
-    }
-}
-*/
